fixing newState type

+++ pkg/wca/IMMNotificationClient.go
@@ -8,7 +8,7 @@ type IMMNotificationClientCallback struct {
 	OnDefaultDeviceChanged func(flow EDataFlow, role ERole, pwstrDeviceId string) error
 	OnDeviceAdded          func(pwstrDeviceId string) error
 	OnDeviceRemoved        func(pwstrDeviceId string) error
-	OnDeviceStateChanged   func(pwstrDeviceId string, dwNewState uint64) error
+	OnDeviceStateChanged   func(pwstrDeviceId string, dwNewState uint32) error
 	OnPropertyValueChanged func(pwstrDeviceId string, key uint64) error
 }
